NXP-19262: make MongoDB query builder code more generic
